[SC-164] moved the file
